requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
paddlepaddle==2.5.2
paddle-inference==2.5.2
langdetect==1.0.9
pydantic==2.4.2
loguru==0.7.2
python-multipart==0.0.6 